+++ prefect_aws/credentials.py
@@ -1,13 +1,25 @@
 """Module handling AWS credentials"""
 
-from typing import Optional
+from enum import Enum
+from typing import Any, Optional, Union
 
 import boto3
-from prefect.blocks.core import Block
+from mypy_boto3_s3 import S3Client
+from prefect.blocks.abstract import CredentialsBlock
 from pydantic import Field, SecretStr
 
+from prefect_aws.client_parameters import AwsClientParameters
 
-class AwsCredentials(Block):
+
+class ClientType(Enum):
+
+    S3 = "s3"
+    ECS = "ecs"
+    BATCH = "batch"
+    SECRETS_MANAGER = "secretsmanager"
+
+
+class AwsCredentials(CredentialsBlock):
     """
     Block used to manage authentication with AWS. AWS authentication is
     handled via the `boto3` module. Refer to the
@@ -27,10 +39,14 @@ class AwsCredentials(Block):
     _block_type_name = "AWS Credentials"
 
     aws_access_key_id: Optional[str] = Field(
-        default=None, description="A specific AWS access key ID."
+        default=None,
+        description="A specific AWS access key ID.",
+        title="AWS Access Key ID",
     )
     aws_secret_access_key: Optional[SecretStr] = Field(
-        default=None, description="A specific AWS secret access key."
+        default=None,
+        description="A specific AWS secret access key.",
+        title="AWS Access Key Secret",
     )
     aws_session_token: Optional[str] = Field(
         default=None,
@@ -38,6 +54,7 @@ class AwsCredentials(Block):
             "The session key for your AWS account. "
             "This is only needed when you are using temporary credentials."
         ),
+        title="AWS Session Token",
     )
     profile_name: Optional[str] = Field(
         default=None, description="The profile to use when creating your session."
@@ -45,6 +62,11 @@ class AwsCredentials(Block):
     region_name: Optional[str] = Field(
         default=None,
         description="The AWS Region where you want to create new connections.",
+    )
+    aws_client_parameters: AwsClientParameters = Field(
+        default_factory=AwsClientParameters,
+        description="Extra parameters to initialize the Client.",
+        title="AWS Client Parameters",
     )
 
     def get_boto3_session(self) -> boto3.Session:
@@ -76,8 +98,38 @@ class AwsCredentials(Block):
             region_name=self.region_name,
         )
 
+    def get_client(self, client_type: Union[str, ClientType]) -> Any:
+        """
+        Helper method to dynamically get a client type.
 
-class MinIOCredentials(Block):
+        Args:
+            client_type: The client's service name.
+
+        Returns:
+            An authenticated client.
+
+        Raises:
+            ValueError: if the client is not supported.
+        """
+        if isinstance(client_type, ClientType):
+            client_type = client_type.value
+
+        client = self.get_boto3_session().client(
+            service_name=client_type, **self.aws_client_parameters.get_params_override()
+        )
+        return client
+
+    def get_s3_client(self) -> S3Client:
+        """
+        Gets an authenticated S3 client.
+
+        Returns:
+            An authenticated S3 client.
+        """
+        return self.get_client(client_type=ClientType.S3)
+
+
+class MinIOCredentials(CredentialsBlock):
     """
     Block used to manage authentication with MinIO. Refer to the
     [MinIO docs](https://docs.min.io/docs/minio-server-configuration-guide.html)
@@ -105,9 +157,18 @@ class MinIOCredentials(Block):
         "for more info about the possible credential configurations."
     )
 
-    minio_root_user: str
-    minio_root_password: SecretStr
-    region_name: Optional[str] = None
+    minio_root_user: str = Field(default=..., description="Admin or root user.")
+    minio_root_password: SecretStr = Field(
+        default=..., description="Admin or root password."
+    )
+    region_name: Optional[str] = Field(
+        default=None,
+        description="The AWS Region where you want to create new connections.",
+    )
+    aws_client_parameters: AwsClientParameters = Field(
+        default_factory=AwsClientParameters,
+        description="Extra parameters to initialize the Client.",
+    )
 
     def get_boto3_session(self) -> boto3.Session:
         """
@@ -140,3 +201,33 @@ class MinIOCredentials(Block):
             aws_secret_access_key=minio_root_password,
             region_name=self.region_name,
         )
+
+    def get_client(self, client_type: Union[str, ClientType]) -> Any:
+        """
+        Helper method to dynamically get a client type.
+
+        Args:
+            client_type: The client's service name.
+
+        Returns:
+            An authenticated client.
+
+        Raises:
+            ValueError: if the client is not supported.
+        """
+        if isinstance(client_type, ClientType):
+            client_type = client_type.value
+
+        client = self.get_boto3_session().client(
+            service_name=client_type, **self.aws_client_parameters.get_params_override()
+        )
+        return client
+
+    def get_s3_client(self) -> S3Client:
+        """
+        Gets an authenticated S3 client.
+
+        Returns:
+            An authenticated S3 client.
+        """
+        return self.get_client(client_type=ClientType.S3)